Moved vite project styles to cra project
