+++ packages/demo-reactjs/package.json
@@ -1,7 +1,9 @@
 {
   "name": "demo-reactjs",
-  "version": "0.1.0",
+  "version": "1.0.0",
+  "license": "UNLICENSED",
   "private": true,
+  "type": "commonjs",
   "dependencies": {
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
@@ -13,8 +15,7 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-scripts": "5.0.1",
-    "typescript": "^4.9.3",
-    "web-vitals": "^2.1.4"
+    "typescript": "^4.9.3"
   },
   "devDependencies": {
     "json-as-xlsx": "*"
@@ -22,8 +23,7 @@
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "test": "react-scripts test"
   },
   "eslintConfig": {
     "extends": [
